Rename new file
