+++ src/standalone_custom/test_shore_power_randomize.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from src.standalone_custom.shore_power_randomize_func import randomize_shore_power
+from src.standalone_custom.shore_power_randomize import randomize_shore_power
 
 def test_randomize_shore_power():
     shore_power_df = {
